# Conflicts:
#	build.gradle
#	src/main/java/com/example/jammoney/financeTerm/entity/UserSavedTerm.java
#	src/main/java/com/example/jammoney/financeTerm/entity/UserTermLearning.java
#	src/main/java/com/example/jammoney/stockApp/kis/dto/StockPriceDto.java
#	src/main/java/com/example/jammoney/stockApp/kis/repository/KisTokenRepository.java
#	src/main/java/com/example/jammoney/stockApp/stock/controller/StockController.java
#	src/main/java/com/example/jammoney/stockApp/stock/entity/Cash.java
#	src/main/java/com/example/jammoney/stockApp/stock/entity/HoldingStock.java
#	src/main/java/com/example/jammoney/stockApp/stock/entity/InterestingStock.java
#	src/main/java/com/example/jammoney/stockApp/stock/entity/Order.java
#	src/main/java/com/example/jammoney/user/entity/User.java

+++ src/main/java/com/example/jammoney/pet/dto/ItemShopResponseDTO.java
@@ -11,6 +11,6 @@ public class ItemShopResponseDTO { //상점에 보여줄 아이템 목록
     private Long itemId;
     private String name; //이름
     private String type;         // ItemType (BACKGROUND, OBJECT)
-    private int price; //가격
+    private long price; //가격
     private String previewUrl;   // 상점 미리보기 이미지
 }
